- Added option mode to GameStates - Key bindings divided into layers - Actions should be updated with respect to new key bindings

+++ include/physics/SolarSystem.h
@@ -18,7 +18,7 @@ protected:
 public:
     SolarSystem();
 
-    int GetBodyCount() {return m_bodies.size();}
+    int GetBodyCount() {return static_cast<int>(m_bodies.size());}
     void Update(float deltaTime);
     void Draw();
     std::vector<std::shared_ptr<CelestialBody>>& GetBodies() { return m_bodies; }
